+++ scripts/build.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+rm -fr dist
+ng build dialog-service
+ng build --prod