+++ scripts/publish-website.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+pushd dist/demo
+surge . dialog-service.surge.sh
+popd