+++ scripts/publish.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+pushd dist/dialog-service
+npm publish
+popd