+++ scripts/serve-prod.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+pushd dist/demo
+python -m SimpleHTTPServer 8000
+popd